+++ src/measurement/area.py
@@ -3,26 +3,16 @@ Created on Nov 20, 2010
 
 @author: patricknevindwyer
 '''
-from measurement import Measurement
+from measurement import Measurement, unitCombiner
+import measurement.length
 
 class Area(Measurement):
     """
     Description of 2 dimensional area. The base unit for Area is the METER.
-    
-    * all measurements are stored in the base unit (in this case METERS)
-    * all measurements are stored as floats
-    * When doing operative math with Numeric arguments, the scale value is first
-        converted into the scaled type. So when doing:
-            
-            (2 * miles) * 5
-        
-        The scaled miles value (2) is converted into the local scaled type (miles)
-        multiplied, and then stored again in the base type. This keeps with the
-        semantic meaning of the current unit, and avoids mathematical gaffes.
-        
-    TODO: fix the op and rop methods to not point to the same code (rop will have different semantics)
     """
     pass
+
+unitCombiner.register_div(Area, measurement.length.Length, measurement.length.meters)
 
 square_meters = Area(
     toBaseUnit = 1.0,